BugFix: revise some features not in older Qt versions + some header includes
QListView::itemAlignment is a new Qt feature (5.12) that I think Qt
Designer sneaked into the spell check language widget when I wasn't
looking.

Also QFile::ExistingOnly is a sub-classed flag from QIODevice that was
introduced in Qt 5.11 and is in fact redundant when used only with
QFile::ReadOnly as that already fails when the file does not exist.

The various other odds and ends are `#include`s that I guess must have
changed because of hidden changes of suchlike in various Qt classes over
various Qt versions.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>