Revise: add more Hunspell dictionary details found on FreeBSD
Also switch to consider the `.aff` files in case there are additional or
supplemental `.dic` files.

Some dictionaries were also found which used a '-' as a separator,
particularly where the language code had a third element. Officially
the language codes should have a lower case first part (language), an upper
case second part (country or large scale classification) and if there is
a third part it should be in "Title" case.  To enable quick look-up the
QMap is populated with all lowercase keys and with all '-' converted to
'_'.

Fix a typo in the Profile Preferences dialogue.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>