Revise: further changes
Change a couple of places so that the text reflects whether the main
dictionary comes with Mudlet or is from the system.

Fix things so that the selected dictionary is in view when the preferences
dialogue is created.

Do not bother to check the return values from Hunspell_add and
Hunspell_remove - they do not seem to be useful.

Improve the dictionary location code to work for using bundled dictionaries
when building in a shadow directory. (I found this whilst building in my
slightly non-standard Windows build environment).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>