BugFix: redo triggerT-type Indication in editor to return a colour coding
We recently removed the foreground colour coding of text entered into the
trigger item text line because it could render some types unreadable with
some themes/styling.  That was not universally popular because the colour
was also a sub-conscious reminder to some users as to the item type.

This PR adds a coloured icon to the trigger type selection `QComboBox` but
- given that it is a little way from the point at which any text is entered
I have also arranged for the coloured (they could also be pictorial icons
if someone wanted to design a set) icons to be duplicated at the start
of the `QLineEdit` where the text is entered.

The colours largely duplicate the colours previously used for those with
text entry and I have chosen a pair of contrasting colours for the two
(colour and prompt) triggers that do not show the text entry widget.

This also properly wraps the trigger type texts in `tr(`...`)` so they
will finally be included in the translation process - from which they
were previously not!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

BugFix: ensure default/unused trig. type icons get copied to text entry

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

Draft: move trigger type QComboBox to LHS of trigger item

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>