Revise: following peer review
Take "Single" from addSingleWordToSet(...)/removeSingleWordFromSet(...) .

Refactor:
* mudlet::prepareProfileDictionary(...)
* mudlet::prepareSharedDictionary()
* mudlet::saveDictionary(...)
so that they use a number of common subroutines.

Change the lua getDictionaryWordList so that its output is sorted in a
case-insensitive manner instead of a case-sensitive manner.

Fix a coding error that left the contents of the dictionary selection
widget in the profile preferences to be unsorted.

Fixed a couple of typos:
* "generatlly" ==> "generally"

Removed "Hunspell" from following members of `TCommandLine`:
* mHunspellSystemDictionarySuggestionsCount
                                      ==> mSystemDictionarySuggestionsCount
* mHunspellProfileDictionarySuggestionsCount
                                     ==> mUserDictionarySuggestionsCount
* mpHunspellSystemSuggestionsList ==> mpSystemSuggestionsList
* mpHunspellProfileSuggestionsList ==> mpUserSuggestionsList

Simplify a couple of `QPair<bool, QString> result` initialisations.

Refactor the common dictionary file location code in
TConsole::setSystemSpellDictionary(...) and
dlgProfilePreferences::initWithHost(...) to be another option for/in the
mudlet::getMudletPath(...) method.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>