Refactor: provide mProfileName to cTelnet, TConsole and TMap classes
This will enable them to establish the identity of their related `Host`
instance without the need to call `Host::getName()` which involves a
mutex in the latter class. This makes things a bit faster and less likely
to cause deadlocks in odd places.

The `Host::setName(const QString&)` method has been revised to ensure that
changes are passed across to the other three classes on change. This will
close https://github.com/Mudlet/Mudlet/issues/2357 - with the exception
that changing the name of the profile whilst it is active will need some
further jiggery-pokery in the HostManager and mudlet classes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>