BugFix: prevent lockup when closing SSL profile
The cause of Issue https://github.com/Mudlet/Mudlet/issues/2288 "Mudlet
freezes when closing secure profile with self-signed certificates on Linux"
seems to be because there are asynchronous signals from the network thread
that arrive when there is not enough of the profile left to handle it.

Putting a call to `cTelnet::disconnectIt()` {was called `disconnect()` but
that seems to overload a non-static variant of `QObject::object(...)` with
all three arguments defaulted} in the `Host` destructor is the primary
suspect in this case!

This commit seems to fix the issue by not allowing the profile to be closed
until it is known to be disconnected from the game server.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>