BugFix: update CMakeFiles.txt for extra lua file
With the addition of ./mudlet-lua/lua/TTSValues.lua to the project in:
https://github.com/Mudlet/Mudlet/commit/34cf653 as part of:
https://github.com/Mudlet/Mudlet/pull/1893
This added a new lua file to the set we provide but the top-level CMake
project file was not updated to account for the extra file so it is now
issuing a warning for each build.  This PR updates that file so it does not
moan anymore about it.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>