Revise: QGridLayouts' order & resolve names in profile_preferences.ui
This file was provoking four warnings on compilation because the same name
was being used in two places for 2x4 = 8 items.

This PR resolves the clashes (mainly in auto-suffixed with "_##") in the
names of some QGridLayout/QVBoxLayouts by changing the suffixes to reflect
the item that they were the layout for.  There was a similar clash with
two unsuffixed items both called "verticalSpacer".

At the same time I have taken the opportunity to reset the order of some
items in `QGridLayout`s in this form so that the row/column values are
in ascending order. Maintaining that order in PRs will help to improve the
signal-to-noise ratio in other git diffs.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>