Revise: update Unicode codepoints width tables to Unicode 11.0
The previously used `./src/wcwidth` `.c` and `.h` only covered up to
Unicode version 5.0 (from 2007-05-26). This commit revises our code to use
an alternative widechar_width.h which public domain licensed and available
as a C++ header:
https://github.com/ridiculousfish/widecharwidth/blob/master/widechar_width.h
{The rest of the project is some python magic that parses the Unicode
source texts and regenerated this file}

This commit imports the file as of 2019/03/04 which has support for
Unicode 11.0 . As Qt versions prior to 5.11.0 only understand Unicode up to
8.0 and some characters changed width (got wider) in the 8.0 to 9.0
transition the code takes account of the run-time version of Qt in use for
those characters so as to try and maintain the width that the Qt Painter
code will expect to use for them.

This is likely to be a stepping stone on the way to solve some issues with
character spacing and highlighting but it seems simple enough to introduce
it immediately whilst I work on the latter...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>