BugFix: fixup various problems and bug found by clang-tidy/clazy
The return of a reference to a temporary value instead of a (return value
optimised?) value was actually provoking conistent fatal crashes for me on
FreeBSD and it wasn't until I fixed things so that I could run clang-tidy
on the code that I found the issue.

As it happens it also pointed out some other bugs involving use of '\'
instead of '/', some erroneous combinations of `&&` and `||` and the
omission of a return value in an actually very probably execution path.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>