Revise: recheck word after adding/removing it from user's dictionary
It is better now at changing the indication between unknown and in the
user's dictionary for a word that is not in the main one - but it is not
perfect, especially if there are additional punctuation marks abutting the
word concerned.

Also added indication of which of the per profile or the shared user
dictionaries is providing the user dictionary suggestions.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>