Enhance: add per profile dictionary capability (#2358)
This is a squash and merge PR and the following is an edited summary
of the individual commit messages that were combined into it:

It is saved at the end of a session and the affix file ('profile.aff') is
maintained to allow for the Hunspell suggestion capability to operate.
The word-list ('profile.dic') file is checked at session start to
accommodate manual editing (adding/removing words) between sessions.

When a word typed on the command line is not found in the main (system
provided on non-Windows OSes) instead of the red (now wavy on Unix,
dotted on macOS) underline a cyan dashed underline is used to show that the
word was found in the profile's own dictionary. Words can also be added
and removed.

Also found out why the wavy / dashed underline was not showing in previous/
initial attempt.

Some lua commands are also added:
* addWordToDictionary((string) word) returns true if the word was
added (and was not already in the dictionary)
* removeWordFromDictionary((string) word) returns true if the word
was present and removed from the dictionary.
* getDictionaryWordList()` returns a (sorted into system locale
case insensitive order) list of ALL words in the profile dictionary.
* `getDictionaryWordList((string) word [,(bool)useProfileDictionary])` do a check
in the specified system dictionary (default) or if a second argument is
supplied and is `true` use the per profile one. Returns `true` if the word
is found present in the dictionary, and false if not.
* spellCheckWord((string) word) [,(bool)useUserDictionary]) returns true if the
word is in the dictionary - uses the main language dictionary as set in the
profile preferences unless a second, optional boolean true argument is
provided then it will use the user's stored word list - either the per profile
or the shared across profiles as set in the profile preferences.
* spellSuggestWord((string) word [,(bool)useUserDictionary]) does the
same sort of suggestions search as is done on the console command line and
returns a list of suggestions from the specified system dictionary
(default) or if a second argument is supplied and is `true` then it will use
the user's stored word list - either the per profile or the shared across
profiles as set in the profile preferences.

Also:
* Arrange to not list supplemental medical dictionaries.
* Improve system dictionary selection by added text for all the
  dictionaries I can identify in my Linux system distribution.
* Reordered some items in the mudlet constructor initialisation list to
  match their placement in the header file - it was helpful to do this as I
  found I needed to add an initialiser for a pointer and wanted to put it
  in the right place.
* Revise: add more Hunspell dictionary details found on FreeBSD

Also switch to consider the `.aff` files in case there are additional or
supplemental `.dic` files.

Some dictionaries were also found which used a '-' as a separator,
particularly where the language code had a third element. Officially
the language codes should have a lower case first part (language), an upper
case second part (country or large scale classification) and if there is
a third part it should be in "Title" case.  To enable quick look-up the
QMap is populated with all lowercase keys and with all '-' converted to
'_'.

Fix things so that the selected dictionary is in view when the preferences
dialogue is created.

Do not bother to check the return values from Hunspell_add and
Hunspell_remove - they do not seem to be useful.

Improve the dictionary location code to work for using bundled dictionaries
when building in a shadow directory. (I found this whilst building in my
slightly non-standard Windows build environment).

Refactor:
* mudlet::prepareProfileDictionary(...)
* mudlet::prepareSharedDictionary()
* mudlet::saveDictionary(...)
so that they use a number of common subroutines.

Change the lua getDictionaryWordList so that its output is sorted in a
case-insensitive manner instead of a case-sensitive manner.

Also handle the case should a main dictionary not be found.

Revise: after peer review disable the option to NOT use a user dictionary

As there is some careful coding in the existing code to prevent unnecessary
loading and unloading of the per profile and shared user dictionaries it
is easier to just prevent the option to disable both of them from being
chosen - at least whilst evaluating things - this will maintain the
presence of the "mEnableUserDictionary = 'yes'" option in new profile
saves.

Revise: recheck word after dict. options changes & clear marks when off

As requested during peer review - this in fact improves upon situation
before the start of adding user dictionaries.

Revise: recheck word after adding/removing it from user's dictionary

It is better now at changing the indication between unknown and in the
user's dictionary for a word that is not in the main one - but it is not
perfect, especially if there are additional punctuation marks abutting the
word concerned.

Also added indication of which of the per profile or the shared user
dictionaries is providing the user dictionary suggestions.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>