Enhance: add some Hunspell dictionaries to our own source code (#2404)
These are for Windows and MacOs users that do not come with Hunspell
dictionaries by default. The files concerned will need to be added to
the distribution bundles.

All but the French files were taken from my Debian Linux distribution as
of 2019/03/11 - it is my belief that these files may be distributed under
the terms of the GPL that we use - some claim (in the .aff) file that they
are (L)GPL licensed and the remainder are under the Mozilla Public Licence
2.0 which is compatible with the GPL.

As I was not sure about the MPL initially I downloaded the French files
from their current distributor https://grammalecte.net/download.php?prj=fr
which is a bit newer (6.3) that the Debian Stretch version (5.7). Further
reading revealed that the MPL 2.0 is suitable for combining into a GPL
project. The variant that I have included is that described as the
"Classical" one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>