Refactor: convert color keybinding QLineEdit to QLabel + other I18n fixes
By making it a label (with a frame) we do not have to worry about giving
the impression as a line-edit set to read-only mode that it is something
that can be directly edited (by default a read-only line edit looks the
same as one that can be typed in). To better make it match in some other
aspects a plain lined boarder is useful.

Whilst editing this it became clear that the default text was not in a
`QObject::tr(...)` wrapper so was not being presented as a text for
translation!

It also became clear that because the modifiers do have some OS dependent
details it is desirable for them to be adjustable by scripts (at least for
the permanent keybindings. This PR adds the means for the lua sub-system
to find out and change the modifiers for the permanent keybinding given
the name of the keybinding.

Tool-tips have been added to the individual check boxes that give guidance
as to cross-platform variation details so that packagers can accommodate
setting / documenting any keybindings for any OS platform.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>