BugFix: prevent crashes when no system spell checker is available/set
This was actually discovered by accident because there was not a default
dictionary specified in the Host constructor initialiser list. However the
principle of handling things safely if there is a problem in setting up
the system spell checker should be adopted even after fixing that point.

This should close https://github.com/Mudlet/Mudlet/issues/2413.

I also found another issue in that using the right click (context menu)
on the command line was feeding the word being checked to the user's
dictionary in the encoding that the system/mudlet provided dictionaty
was using - which need not be the UTF-8 encoding that the former will
always be using - that has been fixed herein as well.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>