Refactor: rationalise TLuaInterpreter access to TConsoles (Part 01)
This revises the lua `selectString(...)` function. As well as selecting the
given instance of the given text in the given window (which is the main
profile `TConsole` if the name is omitted or is the empty string) it also
returns the first character position in the relevant line in the buffer.

There are a couple of other error conditions that can also return an error
(if the selected line is outside of the range of lines within the buffer)
as they are run-time they will now cause a nil plus error message - as will
trying to select an empty string.

I18n note: the character position is still the index into the `QChar`s that
make up the UTF-16 encoded representation of the string - this is
demonstrably WRONG for anything other than ASCII text... especially because
we use UTF-8 in the lua subsystem!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>