Revise: following peer review remove run-time checks for Host's timer map
Instead of checking each time we register a new TTimer with the mudlet
singleton that there is a QMap<QTimer*, TTimer*> entry for it and any
others from the same profile under the Host pointer (and create it if
needed and destroy it when it gets emptied) we now create that QMap entry
in the Host constructor.

This does need an extra condition to not do this for the dummy default_host
as the mudlet class has not been fully constructed when it is instantiated
as adding an empty QMap entity to the QMap<Host*, QMap<QTimer*, TTimer*>>
mudlet::mHostTimerMap causes a seg. fault. on start up.

I have left some debugging Q_ASSERTs in place to check that the removed
checks are no longer needed...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>