Revise: remove colour palettes from editor search area
The palette settings there do not work well when using a dark theme,
removing the previous palette details does not seem to have any other
side effects.

At the same time I have also renamed the editor control:
* (QCheckBox*) dlgTriggerEditor::button_displayAllVariables
                         ==> dlgTriggerEditor::checkBox_displayAllVariables
because it *is* a QCheckBox and not the QPushButton it once was!

I also managed to simplify the widget structure on the left side of the
editor - by removing some container widgets - this should make the layout
work better in extreme cases or if some text elements change their needed
sizes when the GUI language is changed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>