Revise: remove colour palettes from vars_main_area.ui
This is so that there is no colours to interfere with dark Desktop or
Mudlet application themes.

The widgets have been rejigged and a vertical spacer inserted at the bottom
so that in certain situations that may arise in conjunction with other
changes forthcoming in the Editor (get the splitter working properly) the
details area does not swell up to grab the whole right hand side of the
editor such there be no other widgets in the whole of the right side.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>