Revise: remove colour palettes from system message area & dlgTriggerEditor
Refactor the system_message_area.ui file to remove all uses of palettes,
for the places where some colouration is required this can be done by a
much less verbose stylesheet - which was already being used for one
particular widget. Some simplification also removed some unneeded widgets
which will make the form layout work better because there are less nested
layouts. However an additional vertical spacer is needed so that the
whole of the rest of the widget can be made to only be as large as
necessary to display the text should there be NO other widgets on display
in the splitter. This is a change that is also prompted by the desire to
make the notification area be able to expand to show longer messages.

Remove the yellowish colouration of the root items in the treewidget_XXXX
for each item type in the editor so that they are readable whatever colour
theme is in effect.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>