Enhance: get splitter on right side of editor working + translatable texts
Restructure trigger_editor.ui form so that a single splitter on right hand
side is used for all widgets on that side which makes it possible to resize
the various elements in a more flexible way than previously.

Make item help messages be translatable and ensure ones at top of
dlgTriggerEditor.cpp are used (rather than others buried in the add item
of each type) - and allows the space for them to be expanded if needed.

Make column headers in search results treeWidget be translatable.

Revised the code that reported items being enabled/disabled in the editor
so that if there is an error the item is always left disabled instead, it
also reformulates the texts so that they can be translated as a sentence,
with less fragments of text components.

Simplifies the layout of the triggers_main_area.ui and arranges so that
it can expand as necessary for other GUI languages.

Improves sound file selection on trigger UI - including presetting file
selector to previously chosen file or profile home directory (instead of
user home dir) if none and adding a tool button to clear the setting - now
the use of cancel in the file selection dialog does indeed cancel the
process of setting a file rather than clearing a previous setting.

Improve colourising fore/background colour selection UI - including greying
out colour buttons when disabled.

Remove a load of casts from enums to ints by making passed arguments to
some methods to explicitly be those enums {which, after all, is one purpose
of using enums!}

Revise the timer_main_area.ui with less fixed sizes that won't expand as
needed.

Demote several methods from being SLOTS in dlgTriggerEditor class:
* (void) dlgTriggerEditor::slot_action_toggle_active()
                                ==> dlgTriggerEditor::activeToggle_action()
* (void) dlgTriggerEditor::slot_alias_toggle_active()
                                 ==> dlgTriggerEditor::activeToggle_alias()
* (void) dlgTriggerEditor::slot_key_toggle_active()
                                   ==> dlgTriggerEditor::activeToggle_key()
* (void) dlgTriggerEditor::slot_script_toggle_active()
                                ==> dlgTriggerEditor::activeToggle_script()
* (void) dlgTriggerEditor::slot_timer_toggle_active()
                                 ==> dlgTriggerEditor::activeToggle_timer()
* (void) dlgTriggerEditor::slot_trigger_toggle_active()
                               ==> dlgTriggerEditor::activeToggle_trigger()
* (void) dlgTriggerEditor::slot_deleteAlias()
                                       ==> dlgTriggerEditor::delete_alias()
* (void) dlgTriggerEditor::slot_deleteAction()
                                      ==> dlgTriggerEditor::delete_action()
* (void) dlgTriggerEditor::slot_deleteKey()
                                         ==> dlgTriggerEditor::delete_key()
* (void) dlgTriggerEditor::slot_deleteScript()
                                      ==> dlgTriggerEditor::delete_script()
* (void) dlgTriggerEditor::slot_deleteTimer()
                                       ==> dlgTriggerEditor::delete_timer()
* (void) dlgTriggerEditor::slot_deleteTrigger()
                                     ==> dlgTriggerEditor::delete_trigger()
* (void) dlgTriggerEditor::slot_deleteVar()
                                    ==> dlgTriggerEditor::delete_variable()
* (void) dlgTriggerEditor::grab_key_callback()
                                  ==> dlgTriggerEditor::key_grab_callback()

Also removed some that were wrappers around short code fragments that could
be put directly in the single place of use:
* (void) dlgTriggerEditor::slot_addTrigger()
* (void) dlgTriggerEditor::slot_addTriggerGroup()
* (void) dlgTriggerEditor::slot_addTimer();
* (void) dlgTriggerEditor::slot_addTimerGroup()
* (void) dlgTriggerEditor::slot_addAlias()
* (void) dlgTriggerEditor::slot_addAliasGroup()
* (void) dlgTriggerEditor::slot_addScript()
* (void) dlgTriggerEditor::slot_addScriptGroup()
* (void) dlgTriggerEditor::slot_addAction()
* (void) dlgTriggerEditor::slot_addActionGroup()
* (void) dlgTriggerEditor::slot_addKey()
* (void) dlgTriggerEditor::slot_addKeyGroup()
* (void) dlgTriggerEditor::slot_addVar()
* (void) dlgTriggerEditor::slot_addVarGroup()

Renames some `.ui` form components that are manipulated by the code in
the dlgTriggerEdit class so that the element concerned is more clearly
identified and consistent with other names, e.g.:
soundTrigger ==> groupBox_soundTrigger

Correct the too-tip text for the `spinBox_stay_open` from:
"Within how many lines must all condition be true to fire the trigger?"
to:
"How many more line after the one that fired the trigger should be passed
to children?"
as the former text was an erroneous duplicate of the tool-tip for the delta
value for multi-line or AND triggers.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>