Don't reset other profile's timers with resetProfile() (#2257)
This is a squash and merge (of three) commits with the following
edited summary of the commit messages:

This will split up the QMap in the mudlet class that contains all the
QTimer & TTimer paired up in the TTimer class so that when a Host
instance runs its resetProfile() method only the QTimers that are
associated with the TTimers in THAT profile get removed from the mudlet
map.  Previously all of them were nuked in that table.

Cleaned up the TTimer class - renamed the (QTimer*) mpTimer to be
called mpQTimer because in some places we refer to instances of this
class with a mpTimer member variable and sometimes we want to access the
QTimer within and using the same name can lead to confusion.

This also lead me to discover that there was no need to include the latter
in the arguments to mudlet::registerTimer(...) as it is already available
from the previous argument.

I also found out one cause of *bogus* "MUDLET CRITICAL ERROR: trying to
unregister Timer but it is not registered!" messages in that a TTimer is
created as a root item for TTimer items every time
XMLimport::importPackage(...) is run and it is deleted at the end if the
package does not import any timer items. I have added a flag that can be
set in that method so that the (revised) error message is not produced in
*that* case.

Following peer review remove run-time checks for Host's timer map. So
that instead of checking each time we register a new TTimer with the
mudlet singleton that there is a QMap<QTimer*, TTimer*> entry for it
and any others from the same profile under the Host pointer (and create
it if needed and destroy it when it gets emptied) we now create that
QMap entry in the Host constructor.

This does need an extra condition to not do this for the dummy default_host
as the mudlet class has not been fully constructed when it is instantiated
as adding an empty QMap entity to the QMap<Host*, QMap<QTimer*, TTimer*>>
mudlet::mHostTimerMap causes a seg. fault. on start up.

I have left some debugging Q_ASSERTs in place to check that the removed
checks are no longer needed...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>