Enhance: add Event when application Style-sheet set
Using the Lua function `setAppStyleSheet` will likely wipe out any local
changes that Geyser and user scripts have made to things like labels (and
gauges that are based upon them).  It is essential therefore for those to
be aware that this has happened - so this PR provides a
"sysAppStyleSheetChange" event that will be sent to ALL profiles when it
is called from one of them. To allow individual bits of code to trace the
origin of the call (in case it was responsible for it and must therefore
not respond to it in the same way) an optional second (string) argument to
the setAppStyleSheet call will propagate into the event passed by the
system as a second argument (will be an empty string if omitted) and the
source of the event (the profile name) will be included as a third
argument.

At this time I am also producing another PR that will involve the Mudlet
application itself applying an application style sheet - so that that
can be accommodates I will arrange for it too to produce this event type
when it acts - in this case the second argument will likely be one of
"mudlet-theme-dark" or "mudlet-theme-light" and the third will be "system".

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>