Refactor: store the Host name and TTimer Id in QTimer properties
Prevent crashes when using resetProfile() - which gets awkward if the code
to be run that contains the resetProfile() call is itself going to be
deleted by that call!

This eliminates the need for a mudlet class multi-layer QMap:
(QMap<Host*, QMap<QTimer*, TTimer*>>) mHostTimerMap
which otherwise would be consulted to find the needed TTimer instance when
a QTimer fired. This is advantageous because it avoids a whole extra level
of complexity when adding or removing (especially temporary {and thus one-
shot}) TTimers.

It also corrects a confusing issue in that the Lua constructed temp timers
are all initialised with the name "a" - instead they are given a more
useful (for debugging) "newTempTimerWithoutAnId" when constructed in the
TLuaInterpreter class until they are assigned the string form of their Id
number when they are registered.

The clean up code for `TTimer`s was also simplified by switching the
container that held the pointers for `TTimer`s that were to be cleaned up
from a `std::list<TTimer*>` to a `QSet<TTimer*>` which meant it was no
longer necessary to search the whole list when adding a new one to prevent
duplication as a set container inherently prevents duplicates.

I found that it was feasible to remove (bool) TTimer::registerTimer()
and move it's functionality into (bool) TimerUnit::registerTimer(TTimer*) -
this is so there is one less function to ensure is called at the right
time.

Remove unused argument from method:
* (bool) TTimer::canBeUnlocked(TTimer* pChild) ==>
                                             (bool) TTimer::canBeUnlocked()

Also convert a pair of C style casts in Tree.h to a more appropriate C++
style.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>