Enhance: draw RoomId numbers better in 2D Mapper
They now are scaled so that they fit the room symbol square - and are
suppressed if they are too small to show.  Importantly this is done on the
length of the longest number found in the area so that there will not be
truncation - either they will all fit or none will. Also, leading zeros are
used so that they are all the same size...

To distinguish between roomId and any user defined symbols for a room that
might just be a series of numbers, the room Id numbers are also drawn in
bold and with both an under- and an over-line.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>