Revise: merge in changes from the reversion of code this PR was fixing
This commit allows this PR to be merge-able with the development branch
after https://github.com/Mudlet/Mudlet/pull/2486 (which reverted changes
made in https://github.com/Mudlet/Mudlet/pull/2257 as that had flaws that
this PR corrects) was applied.  As such it invalidates the need for the
reversion of the reversion that was proposed in https://github.com/Mudlet/Mudlet/pull/2492.

Conflicts resolved in:
* src/Host.cpp
* src/TTimer.cpp
* src/TTimer.h
* src/XMLimport.cpp
* src/mudlet.cpp
* src/mudlet.h

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>