BugFix: target room selection for speed-walking was inaccurate
It turns out that the area that had to be (double-) clicked on a room to
initiate a speed-walk on the 2D mapper was incorrectly calculated. It was
centred on the QRectF::x() and QRectF::y() which are actually the top-left
corner and not the middle of the room!

Whilst debugging this I made an interesting discovery in that double-
clicking in the area of the coloured arrow that is drawn for an area exit
off to one side of a room with an XY-plane exit to another area will do a
speed-walk to the room in the other area...

... I also found that a chunk of code that would have implemented this for
the case when grid mode was selected - was totally pointless - as it
requires the `(QMap<int, QPoint> T2DMap::mAreaExitsList` member to be
populated and it cannot be as that requires the exits to be examined which
is skipped as they are not considered in grid mode!

Also: fixup some indentation where a past contributor used tabs rather than
spaces for indentation in the area where I was revising the code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>