Revise: avoid changing the objectName of toolbars/docked windows on layout
The objectName is used to identify the item in the opaque data that is
saved to preserve the size and position these items - so changing it is
probably not a good idea.

This PR instead sets and modifies a (dynamic) property "layoutChanged" on
each item instead.

It also revises the code used to reset the property on each item when the
state details are saved, the replacement code is a bit clearer IMHO as
it uses the Qt Java style iterators rather than the STL ones.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>